replace with batch
