+++ manager/pkg/statussyncer/syncers/local_spec_policies_syncer.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"gorm.io/gorm"
+	"gorm.io/gorm/clause"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	statusbundle "github.com/stolostron/multicluster-global-hub/manager/pkg/statussyncer/bundle"
@@ -21,8 +22,8 @@ import (
 	"github.com/stolostron/multicluster-global-hub/pkg/database/models"
 )
 
-// NewLocalSpecPoliciesSyncer creates a new instance of LocalSpecDBSyncer.
-func NewLocalSpecPoliciesSyncer(log logr.Logger) DBSyncer {
+// NewLocalPolicySpecSyncer creates a new instance of LocalSpecDBSyncer.
+func NewLocalPolicySpecSyncer(log logr.Logger) DBSyncer {
 	dbSyncer := &localSpecPoliciesSyncer{
 		log:                             log,
 		createLocalPolicySpecBundleFunc: statusbundle.NewLocalPolicySpecBundle,
@@ -79,82 +80,77 @@ func (syncer *localSpecPoliciesSyncer) handleLocalObjectsBundleWrapper(tableName
 // if the row exists then update it.
 // if the row isn't in the bundle then delete it.
 func (syncer *localSpecPoliciesSyncer) handleLocalObjectsBundle(ctx context.Context, bundle status.Bundle,
-	dbClient postgres.LocalPoliciesStatusDB, schema string, tableName string,
+	dbClient postgres.LocalPoliciesStatusDB, tableSchema string, tableName string,
 ) error {
 	logBundleHandlingMessage(syncer.log, bundle, startBundleHandlingMessage)
 	leafHubName := bundle.GetLeafHubName()
 
 	db := database.GetGorm()
-	policyIdToVersionMapFromDB, err := getPolicyIdToVersionMap(db, schema, tableName, leafHubName)
+	policyIdToVersionMapFromDB, err := getPolicyIdToVersionMap(db, tableSchema, tableName, leafHubName)
 	if err != nil {
-		return fmt.Errorf("failed fetching leaf hub '%s.%s' IDs from db - %w", schema, tableName, err)
+		return fmt.Errorf("failed fetching leaf hub '%s.%s' IDs from db - %w", tableSchema, tableName, err)
 	}
 
-	// https://gorm.io/docs/transactions.html
-	err = db.Transaction(func(tx *gorm.DB) error {
-		for _, object := range bundle.GetObjects() {
-			specificObj, ok := object.(metav1.Object)
-			if !ok {
-				continue
-			}
-			uid := string(specificObj.GetUID())
-			resourceVersionFromDB, objInDB := policyIdToVersionMapFromDB[uid]
+	batchUpsertLocalPolicies := []models.LocalSpecPolicy{}
+	for _, object := range bundle.GetObjects() {
+		specificObj, ok := object.(metav1.Object)
+		if !ok {
+			continue
+		}
+		uid := string(specificObj.GetUID())
+		resourceVersionFromDB, objInDB := policyIdToVersionMapFromDB[uid]
 
-			payload, err := json.Marshal(specificObj)
-			if err != nil {
-				return err
-			}
-
-			// if the row doesn't exist in db then add it.
-			if !objInDB {
-				syncer.log.Info("local policy created", "leafHubName", leafHubName, "policyID", uid)
-				tx.Unscoped().Where(&models.LocalSpecPolicy{
-					PolicyID: uid,
-				}).Delete(&models.LocalSpecPolicy{})
-				tx.Create(&models.LocalSpecPolicy{
-					LeafHubName: leafHubName,
-					Payload:     payload,
-				})
-				continue
-			}
-
-			// remove the handled object from the map
-			delete(policyIdToVersionMapFromDB, uid)
-
-			// update object only if what we got is a different (newer) version of the resource
-			if specificObj.GetResourceVersion() == resourceVersionFromDB {
-				continue
-			}
-			err = tx.Model(&models.LocalSpecPolicy{}).
-				Where(&models.LocalSpecPolicy{
-					PolicyID: uid,
-				}).
-				Updates(models.LocalSpecPolicy{
-					Payload:     payload,
-					LeafHubName: leafHubName,
-				}).Error
-			if err != nil {
-				return err
-			}
+		payload, err := json.Marshal(specificObj)
+		if err != nil {
+			return err
+		}
+		// if the row doesn't exist in db then add it.
+		if !objInDB {
+			batchUpsertLocalPolicies = append(batchUpsertLocalPolicies, models.LocalSpecPolicy{
+				PolicyID:    uid,
+				LeafHubName: leafHubName,
+				Payload:     payload,
+			})
+			continue
 		}
 
-		// delete objects that in the db but were not sent in the bundle (leaf hub sends only living resources).
+		// remove the handled object from the map
+		delete(policyIdToVersionMapFromDB, uid)
+
+		// update object only if what we got is a different (newer) version of the resource
+		if specificObj.GetResourceVersion() == resourceVersionFromDB {
+			continue
+		}
+		batchUpsertLocalPolicies = append(batchUpsertLocalPolicies, models.LocalSpecPolicy{
+			PolicyID:    uid,
+			LeafHubName: leafHubName,
+			Payload:     payload,
+		})
+	}
+
+	err = db.Clauses(clause.OnConflict{
+		UpdateAll: true,
+	}).CreateInBatches(batchUpsertLocalPolicies, 100).Error
+	if err != nil {
+		return err
+	}
+
+	// delete objects that in the db but were not sent in the bundle (leaf hub sends only living resources).
+	// https://gorm.io/docs/transactions.html
+	// https://gorm.io/docs/delete.html#Soft-Delete
+	err = db.Transaction(func(tx *gorm.DB) error {
 		for uid := range policyIdToVersionMapFromDB {
-			// https://gorm.io/docs/delete.html#Soft-Delete
 			err = tx.Where(&models.LocalSpecPolicy{
-				// LeafHubName: leafHubName,
 				PolicyID: uid,
 			}).Delete(&models.LocalSpecPolicy{}).Error
 			if err != nil {
 				return err
 			}
 		}
-
-		// return nil will commit the whole transaction
 		return nil
 	})
 	if err != nil {
-		return fmt.Errorf("failed handling leaf hub '%s.%s' bundle - %w", schema, tableName, err)
+		return fmt.Errorf("failed deleting records from local_spec.policies - %w", err)
 	}
 
 	logBundleHandlingMessage(syncer.log, bundle, finishBundleHandlingMessage)
