+++ manager/pkg/statussyncer/syncers/managed_clusters_syncer.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"gorm.io/gorm"
+	"gorm.io/gorm/clause"
 	clusterv1 "open-cluster-management.io/api/cluster/v1"
 
 	statusbundle "github.com/stolostron/multicluster-global-hub/manager/pkg/statussyncer/bundle"
@@ -77,78 +78,80 @@ func (syncer *ManagedClustersDBSyncer) handleManagedClustersBundle(ctx context.C
 		return fmt.Errorf("failed fetching leaf hub managed clusters from db - %w", err)
 	}
 
-	// https://gorm.io/docs/transactions.html
-	err = db.Transaction(func(tx *gorm.DB) error {
-		for _, object := range bundle.GetObjects() {
-			cluster, ok := object.(*clusterv1.ManagedCluster)
-			if !ok {
-				continue
-			}
-
-			// Initially, if the clusterID is not exist we will skip it until we get it from ClusterClaim
-			clusterId := ""
-			for _, claim := range cluster.Status.ClusterClaims {
-				if claim.Name == "id.k8s.io" {
-					clusterId = claim.Value
-					break
-				}
-			}
-			if clusterId == "" {
-				continue
-			}
-
-			payload, err := json.Marshal(cluster)
-			if err != nil {
-				return err
-			}
-
-			clusterVersionFromDB, exist := clusterIdToVersionMapFromDB[clusterId]
-			if !exist { // cluster not found in the db table
-				syncer.log.Info("cluster created", "leafHubName", leafHubName, "clusterId", clusterId)
-				tx.Unscoped().Where(&models.ManagedCluster{
-					ClusterID: clusterId,
-				}).Delete(&models.ManagedCluster{})
-				tx.Create(&models.ManagedCluster{
-					ClusterID:   clusterId,
-					LeafHubName: leafHubName,
-					Payload:     payload,
-					Error:       database.ErrorNone,
-				})
-				continue
-			}
-
-			// remove the handled object from the map
-			delete(clusterIdToVersionMapFromDB, clusterId)
-
-			if cluster.GetResourceVersion() == clusterVersionFromDB {
-				continue // update cluster in db only if what we got is a different (newer) version of the resource
-			}
-
-			syncer.log.Info("cluster updated", "leafHubName", leafHubName, "clusterId", clusterId)
-			tx.Model(&models.ManagedCluster{}).
-				Where(&models.ManagedCluster{
-					ClusterID: clusterId,
-				}).
-				Updates(models.ManagedCluster{
-					Payload:     payload,
-					LeafHubName: leafHubName,
-				})
+	// batch upsert managed clusters
+	batchUpsertClusters := []models.ManagedCluster{}
+	for _, object := range bundle.GetObjects() {
+		cluster, ok := object.(*clusterv1.ManagedCluster)
+		if !ok {
+			continue
 		}
 
-		// delete objects that in the db but were not sent in the bundle (leaf hub sends only living resources).
+		// Initially, if the clusterID is not exist we will skip it until we get it from ClusterClaim
+		clusterId := ""
+		for _, claim := range cluster.Status.ClusterClaims {
+			if claim.Name == "id.k8s.io" {
+				clusterId = claim.Value
+				break
+			}
+		}
+		if clusterId == "" {
+			continue
+		}
+
+		payload, err := json.Marshal(cluster)
+		if err != nil {
+			return err
+		}
+
+		clusterVersionFromDB, exist := clusterIdToVersionMapFromDB[clusterId]
+		if !exist {
+			batchUpsertClusters = append(batchUpsertClusters, models.ManagedCluster{
+				ClusterID:   clusterId,
+				LeafHubName: leafHubName,
+				Payload:     payload,
+				Error:       database.ErrorNone,
+			})
+			continue
+		}
+
+		// remove the handled object from the map
+		delete(clusterIdToVersionMapFromDB, clusterId)
+
+		if cluster.GetResourceVersion() == clusterVersionFromDB {
+			continue // update cluster in db only if what we got is a different (newer) version of the resource
+		}
+
+		batchUpsertClusters = append(batchUpsertClusters, models.ManagedCluster{
+			ClusterID:   clusterId,
+			LeafHubName: leafHubName,
+			Payload:     payload,
+			Error:       database.ErrorNone,
+		})
+	}
+	err = db.Clauses(clause.OnConflict{
+		UpdateAll: true,
+	}).CreateInBatches(batchUpsertClusters, 100).Error
+	if err != nil {
+		return err
+	}
+
+	// delete objects that in the db but were not sent in the bundle (leaf hub sends only living resources).
+	// https://gorm.io/docs/delete.html#Soft-Delete
+	err = db.Transaction(func(tx *gorm.DB) error {
 		for clusterId := range clusterIdToVersionMapFromDB {
-			// https://gorm.io/docs/delete.html#Soft-Delete
-			tx.Where(&models.ManagedCluster{
+			e := tx.Where(&models.ManagedCluster{
 				LeafHubName: leafHubName,
 				ClusterID:   clusterId,
-			}).Delete(&models.ManagedCluster{})
+			}).Delete(&models.ManagedCluster{}).Error
+			if e != nil {
+				return e
+			}
 		}
-
-		// return nil will commit the whole transaction
 		return nil
 	})
+
 	if err != nil {
-		return fmt.Errorf("failed handling managed clusters bundle - %w", err)
+		return fmt.Errorf("failed deleting managed clusters - %w", err)
 	}
 
 	logBundleHandlingMessage(syncer.log, bundle, finishBundleHandlingMessage)
